+++ config/env.sh
@@ -1,6 +1,7 @@
 
-# listening port number for HTTP 
+# listening port number for HTTP
 export EXPORT_HTTP_PORT=9300
+# export EXPORT_HTTPS_PORT=9302
 # secret key shared across YourLoops Service
 export SESSION_SECRET="SESSION_SECRET config"
 # styx or api url