+++ log.js
@@ -15,16 +15,47 @@
  == BSD2 LICENSE ==
 */
 
-const _ = require('lodash');
-const baseLog = require('bunyan').createLogger({
+import _ from 'lodash';
+import bunyan from 'bunyan';
+
+function ConsoleRawStream() {}
+ConsoleRawStream.prototype.write = function (rec) {
+    const content = JSON.parse(rec);
+
+    // Prefix all JSON keys with 'dbl_' for some k8s parsing log reason.
+    for (const key in content) {
+      const value = content[key];
+      delete content[key];
+
+      const newKey = `dbl_${key}`;
+      if (key === 'level') {
+        content.dbl_level = bunyan.nameFromLevel[value];
+      } else {
+        content[newKey] = value;
+      }
+    }
+
+    const message = JSON.stringify(content);
+    if (rec.level < bunyan.INFO) {
+        console.log(message);
+    } else if (rec.level < bunyan.WARN) {
+        console.info(message);
+    } else if (rec.level < bunyan.ERROR) {
+        console.warn(message);
+    } else {
+        console.error(message);
+    }
+};
+
+const baseLog = bunyan.createLogger({
   name: 'data-export-service',
+  stream: new ConsoleRawStream(),
 });
 
-function createLogger(filename, extraObjects) {
-  const extras = _.cloneDeep(extraObjects == null ? {} : extraObjects);
+function createLogger(filename, extraObjects = {}) {
+  const extras = _.cloneDeep(extraObjects);
   extras.srcFile = filename;
-
   return baseLog.child(extras);
 }
 
-module.exports = createLogger;
+export default createLogger;